refactor(responsive-image): move types
